Connected the database to mysql database and learned how to authenticate and authorize a user using jwt web tokens.

+++ TodoApp/main.py
@@ -4,6 +4,7 @@ from database import engine, SessionLocal
 from sqlalchemy.orm import Session
 from pydantic import BaseModel, Field
 from typing import Optional
+from auth import get_current_user, get_user_exception
 
 app = FastAPI()
 
@@ -28,10 +29,24 @@ class Todo(BaseModel):
 async def read_all(db: Session = Depends(get_db)): # Now the execution of this function depends on the execution of the get_db function so that is why we have used depends 
     return db.query(models.Todos).all()
 
+@app.get("/todos/user")
+async def read_all_by_user(user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
 
-@app.get("/todo/{todo_id}")
-async def read_todo(todo_id: int, db: Session = Depends(get_db)):
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+    # Dictionary is like this {username: ..., id: user_id}
+
+    if user is None:
+        raise get_user_exception()
+
+    return db.query(models.Todos).filter(models.Todos.owner_id == user.get("id")).all()
+
+
+@app.get("/todos/{todo_id}")
+async def read_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+    
+    if user is None:
+        raise get_user_exception()
+
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get("id")).first()
 
     if todo_model is not None:
         return todo_model
@@ -40,12 +55,17 @@ async def read_todo(todo_id: int, db: Session = Depends(get_db)):
 
 
 @app.post("/", status_code=status.HTTP_201_CREATED)
-async def create_todo(todo: Todo, db: Session = Depends(get_db)):
+async def create_todo(todo: Todo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+
+    if user is None:
+        raise get_user_exception()
+
     todo_model = models.Todos()
     todo_model.title = todo.title
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get("id")
 
     db.add(todo_model)
     db.commit()
@@ -54,9 +74,12 @@ async def create_todo(todo: Todo, db: Session = Depends(get_db)):
 
 
 @app.put("/{todo_id}")
-async def update_todo(todo_id:int, todo: Todo, db: Session = Depends(get_db)):
+async def update_todo(todo_id:int, todo: Todo, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
 
-    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+    if user is None:
+        raise get_user_exception()
+
+    todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get("id")).first()
 
     if todo_model is None:
         raise ItemNotFoundException()
@@ -65,6 +88,7 @@ async def update_todo(todo_id:int, todo: Todo, db: Session = Depends(get_db)):
     todo_model.description = todo.description
     todo_model.priority = todo.priority
     todo_model.complete = todo.complete
+    todo_model.owner_id = user.get("id")
 
     db.add(todo_model)
     db.commit()
@@ -73,14 +97,17 @@ async def update_todo(todo_id:int, todo: Todo, db: Session = Depends(get_db)):
 
 
 @app.delete('/{todo_id}')
-async def delete_todo(todo_id: int, db: Session = Depends(get_db)):
+async def delete_todo(todo_id: int, user: dict = Depends(get_current_user), db: Session = Depends(get_db)):
+
+    if user is None:
+        raise get_user_exception()
 
     todo_model = db.query(models.Todos).filter(models.Todos.id == todo_id).first()
     
     if todo_model is None:
         raise ItemNotFoundException()
 
-    db.query(models.Todos).filter(models.Todos.id == todo_id).delete()
+    db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id == user.get("id")).delete()
 
     db.commit()
